Split main page into multiple components
