+++ src/app/home/Overview.tsx
@@ -1,0 +1,12 @@
+import { Button } from "@/components/Button";
+
+export function Overview() {
+    return (
+        <section className="min-h-screen">
+            <h1>Projects</h1>
+            <h1>Blog</h1>
+            <h1>Tools</h1>
+            <Button />
+        </section>
+    );
+}
