+++ src/app/page.tsx
@@ -1,97 +1,11 @@
-import BlurBackground from "@/components/BlurBackground";
-import { serif } from "./fonts";
-import Link from "next/link";
-import { MdEmail } from "react-icons/md";
-import { FaFileAlt, FaGithub, FaLinkedin } from "react-icons/fa";
-import { IconType } from "react-icons";
-import { ScrollHint } from "./ScrollHint";
-import { Button } from "@/components/Button";
+import { HeroPage } from "./home/HeroPage";
+import { Overview } from "./home/Overview";
 
-function Card() {
-    const quickLinks: { href: string, label: string, Icon: IconType, hover: string }[] = [
-        {
-            href: "/resume.pdf",
-            label: "Resume",
-            Icon: FaFileAlt,
-            hover: "group-hover:text-lime-300",
-        },
-        {
-            href: "https://github.com/cursorweb",
-            label: "GitHub",
-            Icon: FaGithub,
-            hover: "group-hover:text-purple-700",
-        },
-        {
-            href: "https://www.linkedin.com/in/jzhang2029/",
-            label: "LinkedIn",
-            Icon: FaLinkedin,
-            hover: "group-hover:text-[#0a66c2]",
-        },
-        {
-            href: "mailto:jzhang1707@outlook.com",
-            label: "Email",
-            Icon: MdEmail,
-            hover: "group-hover:text-amber-600",
-        },
-    ];
-
-    return (
-        <div className="
-            flex gap-2 flex-row md:flex-col
-            rounded-xl
-            backdrop-blur-md
-            bg-white/70 border-black/10
-            dark:bg-stone-800/70
-            p-5
-            md:w-[calc(0.33*748px)]
-            shadow">
-            {quickLinks.map(({ href, label, Icon, hover }, i) => <Link href={href} target="blank_" key={i}>
-                <div className="group p-2 border border-zinc-300 dark:border-zinc-700 rounded hover:bg-gray-500/30 transition">
-                    <Icon className={`${hover} inline-block m-2`} />
-                    <span className="hidden md:inline">{label}</span>
-                </div>
-            </Link>)}
-        </div>
-    );
-}
-
-export default function Main() {
+export default function Home() {
     return (
         <>
-            <main className="min-h-screen inset-shadow">
-                <BlurBackground />
-                <div className="p-10 gap-10 flex flex-col min-h-screen items-center">
-                    {/* landing page horizontal */}
-                    <div className="grow min-w-full min-h-full flex flex-col justify-center gap-10 md:justify-evenly md:flex-row items-center">
-                        {/* hero text */}
-                        <div className="md:text-left text-center flex flex-col justify-center h-full">
-                            <h1 className={`
-                                ${serif.className}
-                                text-transparent
-                                bg-clip-text
-                                bg-linear-45 dark:from-white dark:to-sky-500
-                                from-black to-sky-500
-                                font-bold md:text-9xl text-6xl
-                                mb-5 md:mb-10
-                                pb-5
-                            `}>
-                                <span className="block">Jerry</span>
-                                <span className="block">Zhang</span>
-                            </h1>
-                            <p className={`${serif.className} text-lg`}>Avid programmer, language learner, and musician.</p>
-                        </div>
-                        <Card />
-                    </div>
-                    <ScrollHint />
-                </div>
-            </main>
-            <section className="min-h-screen">
-                <h1>Projects</h1>
-                <h1>Blog</h1>
-                <h1>Tools</h1>
-                <h1>Skills</h1>
-                <Button />
-            </section>
+            <HeroPage />
+            <Overview />
         </>
     );
 }
